Event refactoring and server-side video events - Organize server event handling code - Generate playback event for current playing video
