+++ lib/entities.ts
@@ -2,6 +2,33 @@
  * @file Shared entity definitions between the front-end and backend.
  */
 
+/*============================================================================*
+ * Media API
+ *============================================================================*/
+
+/**
+ * Playback info for the current media file.
+ */
+export type PlaybackInfo = {
+	file_name?: string
+	file_path?: string
+	file_size?: number
+	title?: string
+	paused?: boolean
+	position?: number
+	duration?: number
+	subtitle?: Subtitle
+	loop_a?: number
+	loop_b?: number
+}
+
+/** A single subtitle line. */
+export type Subtitle = {
+	text: string
+	start: number
+	end: number
+}
+
 /**
  * Directory entry for the media file API.
  */
@@ -24,4 +51,13 @@ export type Dir = {
 	name: string
 	path: string
 	list: DirEntry[]
+}
+
+/*============================================================================*
+ * Server events
+ *============================================================================*/
+
+export type EventVideoPlayback = {
+	type: 'video-playback'
+	play: PlaybackInfo | null
 }
